code refactoring - make_chunks moved to util
